Update customerlist delete functionality

+++ Presentation/Scripts/AngularPrj/Controller/CustomerList.Controller.js
@@ -25,6 +25,7 @@
                 else if (value.maritalStatus == 'W') {
                     value.maritalStatus = 'Widowed';
                 }
+                value.OrgDelStatus = value.isDeleted;
             });
 
         }).catch(function (errorResults) {
@@ -34,24 +35,24 @@
     }
 
     //Delete Customer.
+
     $scope.delete = function () {
-        if ($scope.checkIndexs.count != 0) {
-            CustomerService.deleteCustomer($scope.checkIndexs).then(function (results) {
-                $scope.confirmationConfirmation = results;
-            }).catch(function (errorResults) {
-                $scope.error = 'failure loading Customer List', errorResults;
+        $scope.checkIndexs = [];
+        if ($scope.Customers.count != 0) {
+            angular.forEach($scope.Customers, function (value, key) {
+                if (value.isDeleted != value.OrgDelStatus) {
+                     $scope.checkIndexs.push(value);
+                }
             });
-        }
-    }
-    $scope.checkIndexs = [];
-    $scope.checkIndex = function (customers) {
-        if ($scope.checkIndexs.indexOf(customers) === -1) {
-            $scope.checkIndexs.push(customers);
-            var a = $scope.checkIndexs.indexOf(customers);
-        }
-        else {
-            $scope.checkIndexs.splice($scope.checkIndexs.indexOf(customers), 1)
-            $scope.checkIndexs.push(customers);
+            if ($scope.checkIndexs.count != 0) {
+                {
+                    CustomerService.deleteCustomer($scope.checkIndexs).then(function (results) {
+                        $scope.confirmationConfirmation = results;
+                    }).catch(function (errorResults) {
+                        $scope.error = 'failure loading Customer List', errorResults;
+                    });
+                }
+            }
         }
     };
 }]);
